refactor: add interface for bucket storage and modify docker compose file
