+++ docker-compose.yml
@@ -8,6 +8,8 @@ services:
       - "8081:8081"
     expose:
       - 8081
+    env_file:
+      - .env
 
   postgres:
     image: postgres:alpine
