+++ src/main/kotlin/ingsis/group12/snippetoperations/snippet/service/SnippetServiceImpl.kt
@@ -1,5 +1,6 @@
 package ingsis.group12.snippetoperations.snippet.service
 
+import ingsis.group12.snippetoperations.bucket.ObjectStoreService
 import ingsis.group12.snippetoperations.exception.SnippetCreationError
 import ingsis.group12.snippetoperations.exception.SnippetDeleteError
 import ingsis.group12.snippetoperations.exception.SnippetNotFoundError
@@ -8,12 +9,6 @@ import ingsis.group12.snippetoperations.snippet.input.SnippetInput
 import ingsis.group12.snippetoperations.snippet.model.Snippet
 import ingsis.group12.snippetoperations.snippet.repository.SnippetRepository
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.http.HttpEntity
-import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpMethod
-import org.springframework.http.MediaType
-import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
 import org.springframework.web.client.RestTemplate
 import java.util.UUID
@@ -21,8 +16,7 @@ import java.util.UUID
 @Service
 class SnippetServiceImpl(
     private val snippetRepository: SnippetRepository,
-    @Value
-    ("\${bucket.url}") private val bucketUrl: String,
+    private val objectStoreService: ObjectStoreService,
 ) : SnippetService {
     @Autowired
     private lateinit var restTemplate: RestTemplate
@@ -32,8 +26,7 @@ class SnippetServiceImpl(
         userId: String,
     ): SnippetDTO {
         val snippetId = UUID.randomUUID()
-        val url = "$bucketUrl/$snippetId"
-        val response = addSnippetIntoBucket(snippetInput, url)
+        val response = objectStoreService.create(snippetInput.content, snippetId)
         if (response.statusCode.is2xxSuccessful) {
             return saveSnippet(snippetInput, snippetId)
         }
@@ -44,7 +37,7 @@ class SnippetServiceImpl(
         val result = snippetRepository.findById(snippetId)
         if (result.isPresent) {
             val snippet = result.get()
-            val content = getSnippetFromBucket(snippetId).body!!
+            val content = objectStoreService.get(snippetId).body!!
             return SnippetDTO(
                 snippetId,
                 snippet.name!!,
@@ -59,7 +52,7 @@ class SnippetServiceImpl(
     override fun deleteSnippetById(snippetId: UUID): String {
         val result = snippetRepository.findById(snippetId)
         if (result.isPresent) {
-            val response = deleteSnippetFromBucket(snippetId)
+            val response = objectStoreService.delete(snippetId)
             if (response.statusCode.is2xxSuccessful) {
                 snippetRepository.deleteById(snippetId)
                 return "Snippet deleted with id $snippetId"
@@ -89,25 +82,5 @@ class SnippetServiceImpl(
             snippetInput.language,
             snippetInput.extension,
         )
-    }
-
-    private fun getSnippetFromBucket(snippetId: UUID): ResponseEntity<String> {
-        val url = "$bucketUrl/$snippetId"
-        return restTemplate.getForEntity(url, String::class.java)
-    }
-
-    private fun addSnippetIntoBucket(
-        snippet: SnippetInput,
-        url: String,
-    ): ResponseEntity<String> {
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        val entity = HttpEntity(snippet.content, headers)
-        return restTemplate.postForEntity(url, entity, String::class.java)
-    }
-
-    private fun deleteSnippetFromBucket(snippetId: UUID): ResponseEntity<String> {
-        val url = "$bucketUrl/$snippetId"
-        return restTemplate.exchange(url, HttpMethod.DELETE, null, String::class.java)
     }
 }
